% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{characterDict}
\alias{characterDict}
\title{Character or factor dataframe change to integer matrix.}
\usage{
characterDict(charX, levels)
}
\arguments{
\item{charX}{character dataframe or factor dataframe}

\item{levels, }{the desired levels for each character vector}
}
\value{
list intX the integer matrix x, with crosspoding levels.
}
\description{
Character dataframe, for each column, use \code{characterToInteger()} with unique value from character vector
}
