% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{characterToInteger}
\alias{characterToInteger}
\title{Character vector to Integer vector use Rcpp}
\usage{
characterToInteger(x, levels)
}
\arguments{
\item{x}{original character vector}

\item{levels}{the unique value form x, where the index is the output integer vector value}
}
\value{
integer vector
}
\description{
Character vector to Integer vector use Rcpp
}
