% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{characterToInterger}
\alias{characterToInterger}
\title{Charater vector to Interger vector use Rcpp}
\usage{
characterToInterger(x, levels)
}
\arguments{
\item{x}{original character vector}

\item{levels}{the unique value form x, where the index is the output integer vector value}
}
\value{
interger vector
}
\description{
Charater vector to Interger vector use Rcpp
}
