% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check}
\alias{check}
\title{Check the condition and print message}
\usage{
check(logicstatus, message)
}
\arguments{
\item{logicstatus}{whehter it is true}

\item{message}{message want to print}
}
\description{
Check the condition and print message
}
