% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dataFramToNumeric}
\alias{dataFramToNumeric}
\title{R dataframe to numeric vector}
\usage{
dataFramToNumeric(numX)
}
\arguments{
\item{numX}{numerical dataframe}
}
\value{
numerical matrix
}
\description{
Function will change the original numerical dataframe to a numerical matrix, where missing value will change to Inf.
}
