% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlpha.R
\name{.getAlpha}
\alias{.getAlpha}
\title{Get bootstarp alpha based on previous result.}
\usage{
.getAlpha(bootAlpha)
}
\arguments{
\item{bootAlpha}{bootstrap alpha matrix}
}
\description{
Get bootstarp alpha based on previous result.
}
