% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsguide.R
\name{mrs.pure}
\alias{mrs.pure}
\title{MrSGUIDE internal function call}
\usage{
mrs.pure(
  nX,
  cX,
  Y,
  Trt,
  splitIndex,
  fitIndex,
  holdIndex,
  bk,
  maxDepth = 10L,
  minTrt = 5L,
  minData = max(c(minTrt * maxDepth, NROW(Y)/20)),
  batchNum = 1L,
  CVFolds = 10L,
  CVSE = 0.5,
  bootNum = 500L,
  alpha = 0.05,
  faster = FALSE,
  display = FALSE,
  varName = NULL,
  treeName = paste0("tree_", format(Sys.time(), "\%m\%d"), ".yaml"),
  nodeName = paste0("node_", format(Sys.time(), "\%m\%d"), ".txt"),
  bootName = paste0("boot_", format(Sys.time(), "\%m\%d"), ".txt"),
  impName = paste0("imp_", format(Sys.time(), "\%m\%d"), ".txt")
)
}
\arguments{
\item{nX}{numerical X matrix}

\item{cX}{categorical X matrix}

\item{Y}{outcome Y matrix}

\item{Trt}{treatment vector}

\item{splitIndex}{variable used for split}

\item{fitIndex}{variables can be used for fit}

\item{holdIndex}{variable must include in the fitting model}

\item{bk}{maximal number of variables used in the outcome model for prognostic control}

\item{maxDepth}{maximal depth}

\item{minTrt}{minimum treatment and placebo sample in each node}

\item{minData}{minimum sample in each node}

\item{batchNum}{related with exhaustive search for numerical split variable}

\item{CVFolds}{cross validataion times}

\item{CVSE}{cross validation SE}

\item{bootNum}{bootstrap number}

\item{alpha}{desire alpha levels for confidence interval with respect to treatment parameters}

\item{faster}{related with tree split searching}

\item{display}{Whether display tree in the end}

\item{varName}{variable names}

\item{treeName}{yaml file for save the tree}

\item{nodeName}{file same for each node}

\item{bootName}{boostrap calibrate alpha}

\item{impName}{important variable file name}
}
\description{
Multiple responses subgroup identification use GUIDE gi option, calling C++
}
\details{
Pure GUIDE function in subgroup identification use GI method.

Perform GUIDE gi option for subgroup identification
}
\author{
Peigen Zhou
}
