% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\docType{package}
\name{MrS}
\alias{MrS}
\alias{MrSFit}
\alias{MrS-package}
\title{MrS}
\usage{
MrSFit(dataframe, role, bestK = 1, maxDepth = 5, minTrt = 5,
  minData = max(c(minTrt * maxDepth, NROW(Y)/20)), batchNum = 1L,
  CVFolds = 10L, CVSE = 0, bootNum = 0L, alpha = 0.05,
  faster = FALSE, display = FALSE, treeName = paste0("tree_",
  format(Sys.time(), "\%m\%d"), ".yaml"), nodeName = paste0("node_",
  format(Sys.time(), "\%m\%d"), ".txt"), bootName = paste0("boot_",
  format(Sys.time(), "\%m\%d"), ".txt"), seed = 123, writeTo = FALSE)
}
\arguments{
\item{dataframe}{train data frame}

\item{role}{role follows GUIDE role}

\item{bestK}{number of covariates in the regression model}

\item{maxDepth}{maximum tree depth}

\item{minTrt}{minimum treatment and placebo sample in each node}

\item{minData}{minimum sample in each node}

\item{CVFolds}{cross validataion times}

\item{CVSE}{cross validation SE}

\item{bootNum}{bootstrap number}

\item{alpha}{desire alpha levels for confidence interval with respect to treatment parameters}

\item{faster}{related with tree split searching}

\item{display}{Whether display tree in the end}

\item{treeName}{yaml file for save the tree}

\item{nodeName}{file same for each node}

\item{bootName}{boostrap calibrate alpha}

\item{seed}{random seed}

\item{writeTo}{debug option}

\item{bathNum}{related with exhaustive search for numerical split variable}
}
\description{
Multiple responses subgroup identification use GUIDE gi option

Mr.S identification use GUIDE algorithm
}
\details{
MrS Multiple responses subgroup identification
}
\author{
Peigen Zhou

Peigen Zhou <pzhou9@wisc.edu>
}
