% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsguide.R
\name{MrSFit}
\alias{MrSFit}
\title{MrSGUIDE fit function}
\usage{
MrSFit(
  dataframe,
  role,
  bestK = 1,
  bootNum = 0L,
  alpha = 0.05,
  maxDepth = 5,
  minTrt = 5,
  minData = max(c(minTrt * maxDepth, NROW(Y)/20)),
  batchNum = 1L,
  CVFolds = 10L,
  CVSE = 0,
  faster = FALSE,
  display = FALSE,
  treeName = paste0("tree_", format(Sys.time(), "\%m\%d"), ".yaml"),
  nodeName = paste0("node_", format(Sys.time(), "\%m\%d"), ".txt"),
  bootName = paste0("boot_", format(Sys.time(), "\%m\%d"), ".txt"),
  impName = paste0("imp_", format(Sys.time(), "\%m\%d"), ".txt"),
  writeTo = FALSE,
  remove = TRUE
)
}
\arguments{
\item{dataframe}{train data frame}

\item{role}{role follows GUIDE role}

\item{bestK}{number of covariates in the regression model}

\item{bootNum}{bootstrap number}

\item{alpha}{desire alpha levels for confidence interval with respect to treatment parameters}

\item{maxDepth}{maximum tree depth}

\item{minTrt}{minimum treatment and placebo sample in each node}

\item{minData}{minimum sample in each node}

\item{batchNum}{related with exhaustive search for numerical split variable}

\item{CVFolds}{cross validation times}

\item{CVSE}{cross validation SE}

\item{faster}{related with tree split searching}

\item{display}{Whether display tree in the end}

\item{treeName}{yaml file for save the tree}

\item{nodeName}{file same for each node}

\item{bootName}{file save bootstrap calibrate alpha}

\item{impName}{important variable file name}

\item{writeTo}{debug option reserve for author...}

\item{remove}{whether to remove extra files}
}
\description{
Multiple responses subgroup identification with GUIDE gi option
}
\details{
MrSGUIDE Multiple responses subgroup identification
}
\examples{
library(MrSGUIDE)
set.seed(1234)

N = 200
np = 3

numX <- matrix(rnorm(N * np), N, np) ## numerical features
gender <- sample(c('Male', 'Female'), N, replace = TRUE)
country <- sample(c('US', 'UK', 'China', 'Japan'), N, replace = TRUE)

z <- sample(c(0, 1), N, replace = TRUE) # Binary treatment assignment

y1 <- numX[, 1] + 1 * z * (gender == 'Female') + rnorm(N)
y2 <- numX[, 2] + 2 * z * (gender == 'Female') + rnorm(N)

train <- data.frame(numX, gender, country, z, y1, y2)
role <- c(rep('n', 3), 'c', 'c', 'r', 'd', 'd')

mrsobj <- MrSFit(dataframe = train, role = role)
printTree(mrsobj)

}
\author{
Peigen Zhou
}
