% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTree}
\alias{plotTree}
\title{Plot MrS regression tree}
\usage{
plotTree(
  mrsobj,
  digits = 3,
  height = "600px",
  width = "100\%",
  nodefontSize = 16,
  edgefontSize = 14,
  minNodeSize = 15,
  maxNodeSize = 30,
  nodeFixed = FALSE,
  edgeColor = "#8181F7",
  highlightNearest = list(enabled = TRUE, degree = list(from = 50000, to = 0), hover =
    FALSE, algorithm = "hierarchical"),
  collapse = list(enabled = FALSE, fit = TRUE, resetHighlight = TRUE, clusterOptions =
    list(fixed = TRUE, physics = FALSE)),
  alphaInd = 3
)
}
\arguments{
\item{mrsobj}{MrS object}

\item{digits}{digits for split threshold}

\item{height}{figure height}

\item{width}{figure width}

\item{nodefontSize}{node font size}

\item{edgefontSize}{edge font size}

\item{minNodeSize}{minimal node size}

\item{maxNodeSize}{maximum node size}

\item{nodeFixed}{whether you can drag node}

\item{edgeColor}{edge color}

\item{highlightNearest}{choose node will highligh nearby}

\item{collapse}{list, collapse or not using double click on a node}

\item{alphaInd}{1 is original alpha, 2 is individual level alpha, 3 is overall alpha}
}
\description{
Plot MrS regression tree
}
