% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictTree}
\alias{predictTree}
\title{Predict the node id of MrS regression tree}
\usage{
predictTree(mrsobj, dataframe, type = "nodeid")
}
\arguments{
\item{mrsobj}{MrS object}

\item{dataframe}{data used for prediction}

\item{type}{node id}
}
\description{
Predict the node id of MrS regression tree
}
