% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.guide}
\alias{print.guide}
\title{Print fitted regression tree}
\usage{
\method{print}{guide}(mrsobj, digits = 3, details = FALSE, ...)
}
\arguments{
\item{mrsobj}{MrS object}

\item{digits}{digits pass to coefficent}

\item{details}{whether to print fitting details}
}
\description{
Print fitted regression tree
}
