% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_node}
\alias{print_node}
\title{print node information}
\usage{
print_node(
  node,
  depth = 0,
  digits = 3,
  long = TRUE,
  yName,
  trtName,
  tlevels,
  clevels,
  ...
)
}
\arguments{
\item{node}{node object}

\item{depth}{current depth}

\item{digits}{digit print}

\item{long}{default=TRUE useless}

\item{yName}{outcomes names}

\item{trtName}{treatment name}

\item{tlevels}{treatment levels}

\item{clevels}{categorical variables' levels}

\item{...}{pass to cat}
}
\description{
print node information
}
