% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printTree}
\alias{printTree}
\title{Print fitted regression tree}
\usage{
printTree(mrsobj, digits = 3, details = TRUE, ...)
}
\arguments{
\item{mrsobj}{MrS object}

\item{digits}{digits pass to coefficent}

\item{details}{whether to print fitting details}

\item{...}{parameter pass to \code{print_node}}
}
\description{
Print fitted regression tree
}
