% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{writeTex}
\alias{writeTex}
\title{Write Latex file for GUIDE regression Tree}
\usage{
writeTex(mrsobj, file, digits = 3, ...)
}
\arguments{
\item{mrsobj}{MrS object}

\item{file}{latex filename}

\item{digits}{digits pass to coefficient}

\item{...}{parameters pass to cat function}
}
\description{
Write Latex file for GUIDE regression Tree
}
