% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GiStepWisePure}
\alias{GiStepWisePure}
\title{MrSGUIDE stepwise}
\usage{
GiStepWisePure(
  numX,
  catX,
  Y,
  trt,
  splitIndex,
  fitIndex,
  holdIndex,
  bestK,
  maxDepth,
  minData,
  minTrt,
  batchNum,
  CVFold,
  CVSE,
  bootNum,
  alpha,
  faster,
  display,
  varName,
  treeName,
  nodeName,
  bootName,
  impName
)
}
\arguments{
\item{numX}{numerical X matrix}

\item{catX}{categorical X matrix}

\item{Y}{outcome Y matrix}

\item{trt}{treatment vector}

\item{splitIndex}{variable used for split}

\item{fitIndex}{variables can be used for fit}

\item{holdIndex}{variable must include in the fitting model}

\item{bestK}{maximal number of variables used in the outcome model for prognostic control}

\item{maxDepth}{maximal depth}

\item{minData}{minimum sample in each node}

\item{minTrt}{minimum treatment and placebo sample in each node}

\item{batchNum}{related with exhaustive search for numerical split variable}

\item{CVFold}{cross validation times}

\item{CVSE}{cross validation SE}

\item{bootNum}{bootstrap number}

\item{alpha}{desire alpha levels for confidence interval with respect to treatment parameters}

\item{faster}{related with tree split searching}

\item{display}{Whether display tree in the end}

\item{varName}{variable names}

\item{treeName}{yaml file for save the tree}

\item{nodeName}{file same for each node}

\item{bootName}{file save bootstrap calibrate alpha}

\item{impName}{important variable file name}
}
\description{
MrSGUIDE C++ function
}
\author{
Peigen Zhou
}
