% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs.R
\name{mrs.pure}
\alias{mrs.pure}
\title{Pure GUIDE function in subgroup identification use GI method.}
\usage{
mrs.pure(
  nX,
  cX,
  Y,
  Trt,
  splitIndex,
  fitIndex,
  holdIndex,
  bk,
  maxDepth = 10L,
  minTrt = 5L,
  minData = max(c(minTrt * maxDepth, NROW(Y)/20)),
  batchNum = 1L,
  CVFolds = 10L,
  CVSE = 0.5,
  bootNum = 500L,
  alpha = 0.05,
  faster = FALSE,
  display = FALSE,
  varName = NULL,
  treeName = paste0("tree_", format(Sys.time(), "\%m\%d"), ".yaml"),
  nodeName = paste0("node_", format(Sys.time(), "\%m\%d"), ".txt"),
  bootName = paste0("boot_", format(Sys.time(), "\%m\%d"), ".txt"),
  impName = paste0("imp_", format(Sys.time(), "\%m\%d"), ".txt"),
  seed = 123
)
}
\description{
Perform GUIDE gi option for subgroup identification
}
